refactored API calls
Tasks API calls are in the context now
