+++ src/components/UI/TaskInputField.jsx
@@ -3,14 +3,14 @@ import './TaskInputField.css'
 import AddIcon from '../icons/AddIcon';
 import TaskMenu from './TaskMenu';
 
-function TaskInputField({addTask, handleTaskMenuToggle, isTaskMenuOpen}) {
+function TaskInputField({handleTaskMenuToggle, isTaskMenuOpen}) {
 
   return (
     <>
       <div className='ModifyTaskContainer'>
         {
           isTaskMenuOpen ? (
-            <TaskMenu onClose={handleTaskMenuToggle} addTask={addTask}/>
+            <TaskMenu onClose={handleTaskMenuToggle}/>
           ) : (
             <button 
               onClick={handleTaskMenuToggle}
